requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
gunicorn==19.6.0
MySQL-python==1.2.5
psycopg2==2.6.2
pystache==0.5.4
static==1.1.1
static3==0.7.0
wsgiref==0.1.2
whitenoise==3.2.2
requests==2.12.3
python-geoip==1.2
python-geoip-geolite2==2015.0303
python-whois==0.6.3
